requirements.txt
absl-py==1.3.0
Babel==2.12.1
cffi==1.15.1
future==0.18.3
google-auth-oauthlib==0.4.6
i18n==0.2
kociemba==1.2.1
numpy==1.22.0
oauthlib==3.2.2
opencv-python==4.6.0.66
Pillow==9.3.0
py==1.11.0
pycparser==2.21
pydocstyle==5.1.1
pyglet==2.0.7
PyOpenGL==3.1.6
python-i18n==0.3.9
requests-oauthlib==1.3.1
snowballstemmer==2.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.13.0
torchinfo==1.7.1
torchvision==0.14.0
